ALSA: ctl: add serializer/deserializer of 'elem_value' structure for i386 ABI compatibility
A previous commit adds serializer/deserializer of 'snd_ctl_elem_value'
structure for System V ABIs to modern 32 bit architectures. However, i386
is an exception because it has 4 bytes alignment for double-word machine
type. This brings different 'internal padding' to the structure from the
one on ABIs for modern 32 bit architectures.

This commit adds a pair of serializer and deserializer to convert data
between different layout of the structure on the ABIs. To describe
layout of the structure for i386 ABI on LP64 ABI, bitfield is used for
the internal padding, and 'packed' attribute is used as a hint for
compilers to native padding convention.

The serializer/deserializer are useless on non-Intel architectures. For
this case, '__maybe_unused' is added as a hint to compilers for
optimization.

Reference: System V application binary interface Intel386 architecture
           processor supplenent, draft copy of forth edition (1997,
           The Santa Cruz Operation, Inc. and AT&T)
Reference: System V application binary interface AMD64 architecture
           processor supplement (with LP64 and ILP32 programming models)
           draft version 0.99.8 (2017, H.J. Lu, Michael Matz, Milind
           Girkar, Jan Hubicka, Andreas Jaeger and Mark Mitchell)