ALSA: ctl: add a helper function to allocate for ELEM_LIST request
When investigating compatibility layer, several functions for native ABI
are called by the layer. A 'snd_ctl_elem_list()' is such a function,
while it cannot receive a second argument for data on kernel space. This
will bring a future inconvenience in integration of the layer.

This commit adds a helper function to allocate memory object on kernel
space for the argument.