ALSA: ctl: introduce local framework to handle ioctl requests
In ALSA control core, behaviour of a typical handler for ioctl request can
be described as a set of below tasks:
1.allocate memory object on kernel space
2.copy data of structure from user space to the memory object
3.process data on the memory object according to request type
4.copy data from the memory object to user space

At present, the task of 1, 2 and 4 is programmed to be ad-hoc codes for
each of handler. This is due to differences of actual implementation for
the tasks, while it can be properly represented by structured way.

This commit adds a consistent way to program the above tasks by introduce
a local framework. Entries for each handler are described by:
 * cmd: a command of ioctl(2). The 'size' field is used for task 1.
 * func: for task 3.

In this framework, task 2 and 4 are done by common code.

Unfortunately, current driver programming interface for TLV feature of
ALSA control interface is tight-coupled with a pointer on user space.
To pass the pointer, an unique entry is added for handlers of ioctl
requests for the TLV feature.